v.io/x/lib/gosh: add support for windows and use ErrorDepth correctly.
